Update to 2020b
tzcode.patch doesn't apply cleanly, so I just removed it,
and it worked well on my machine without it.

Change nameToLabel to replace + to ' and not empty string,
this is due Etc-GMT+0 and Etc-GMT0 timezones, which would map to the
same otherwise.

Make tzdata to not `wget` if files are locally.
(My internet is sometimes flaky)